Add as_ argument and handling to transform functions
Most of the available transform functions have a required property `as`
in the vega-lite specification. Since `as` is a special keyword in
python it is not possible to pass as as a normal keyword argument. This
adds an `as_` keyword to these functions in addition to error handling
to propagate this property to the created vega-lite.